# [CI] Add mteb testing to test the accuracy of the embedding model