@noooop 2025-04-25
@DarkLight1337 

Delete the benchmarks/eval/test_mteb.py finally, or change it into a more general testing script.