@noooop 2025-04-28
@DarkLight1337 

I tested more models, and most of them showed strong numerical stability ( <1e-4 ), even better than I imagined. 

The score differences between different models are also quite noticeable ( >1e-3).

This makes mteb STS12 a great embedding model test set.