@noooop 2025-04-30
Thanks for reviewing. 

By carefully studying the code, I indeed learned a lot of weird stuff.