@DarkLight1337 2025-05-01
The failing doc build seems related to this PR. Maybe because you changed the dependencies