@noooop 2025-05-01
> The failing doc build seems related to this PR. Maybe because you changed the dependencies

I only modified the dependencies for test, not for doc build.