@noooop 2025-05-01
[2025-04-30T20:26:36Z] awscli 1.35.23 requires docutils<0.17,>=0.10, but you have docutils 0.21.2 which is incompatible.


https://pypi.org/project/awscli/#history

Maybe a new version of awscli was released.



1.40.5 May 1, 2025

1.40.4 Apr 30, 2025


1.40.3 Apr 29, 2025