@noooop 2025-05-01
I check why such an old version is being used.

awscli==1.35.23
    # via -r requirements/test.in

docutils==0.16
    # via awscli