@noooop 2025-05-01
@DarkLight1337 

I think I can't solve it.

in docs.txt
sphinx 8.2.3 depends on docutils<0.22 and >=0.20

in test.txt (Generated from test.in)
awscli  requires docutils>=0.18.1,<=0.19

https://github.com/aws/aws-cli/blob/6075e33bcdf5d314c092e780d92ea0b63b5ccf4c/setup.py#L28

But wait, why does the docs  environment need to install test dependencies?