@DarkLight1337 2025-05-01
It seems that `awscli` isn't used anywhere in vLLM. Let me open another PR to remove it first