@noooop 2025-05-01
Is pulling a model from s3 dependent on awscli?