@noooop 2025-05-01
I think the issue is not with awscli; I need to set up a local environment for the docs and check it out.