@DarkLight1337 2025-05-01
> Is pulling a model from s3 dependent on awscli?

No, we use `boto3` for that