@noooop 2025-05-01
mainly, mteb depends on "datasets>=2.19.0,<3.0.0" is too old