@noooop 2025-05-01
wtf

```
Uninstalling mteb-1.38.3:                                                                                                                                                                     Would remove:                                                                                                                                                                                 /share/anaconda3/bin/mteb                                                                                                                                                                   /share/anaconda3/lib/python3.12/site-packages/docs/*                                                                                                                                        /share/anaconda3/lib/python3.12/site-packages/mteb-1.38.3.dist-info/*                                                                                                                       /share/anaconda3/lib/python3.12/site-packages/mteb/*     

```

mteb installed a library named docs for python, which caused:

```
Traceback                                                                                                                                                                                   =========                                                                                                                                                                                                                                                                                                                                                                                     File "/share/PycharmProjects/noooop_vllm/docs/source/conf.py", line 137, in setup                                                                                                             from docs.source.generate_examples import generate_examples                                                                                                                             ModuleNotFoundError: No module named 'docs.source'     
```