@DarkLight1337 2025-05-01
Wouldn't this also remove the "correct" docs installation?