@noooop 2025-05-02
I will remove mteb docs from Docker image instead of deleting them from the build docs code, temporarily.

 I will submit an issue to mteb and delete this temporary code once a new version is released.