@noooop 2025-05-20
@DarkLight1337 

The failed tests seem to be unrelated to this PR, so we can finally merge the mteb tests.