@DarkLight1337 2025-05-20
Unblocked embedding models test, see if it passes