@noooop 2025-05-20
Thanks to the mteb community and especially @Samoed for help.